pyproject.toml
[tool.poetry]
name = "polymath-core"
version = "0.1.0.dev54"
description = ""
authors = ["Grady Zhuo <gradyzhuo@gmail.com>"]

packages = [
    { include = "polymath" }
]

[tool.poetry.dependencies]
python = ">=3.8, <3.10"
envparse = "^0.2.0"
pyyaml = "^6.0"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.4.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
